+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fixed issue with `ArcadeSolver` where stacked/overlapped tiles would double solve the position of the collider for the same overlap
+- Fixed issue where initial Actor anchors set in the constructor were not being set in the graphics component
 - EventDispatcher
   - `EventDispatcher` - doesn't require the target object. The context of `this` is not tampered anymore.
 - Pointers